+++ doc/en/conf.py
@@ -14,7 +14,7 @@ import os
 import sys
 sys.path.insert(0, os.path.join(
     # os.path.dirname(os.path.abspath(__file__)),
-    os.path.join('..', '../milvus')
+    os.path.join('.')
 ))
 
 
@@ -22,10 +22,10 @@ sys.path.insert(0, os.path.join(
 
 project = 'pymilvus'
 copyright = '2019, Zilliz'
-author = 'YangXuan'
+author = 'Yinghao.Zou'
 
 # The full version, including alpha/beta/rc tags
-release = '0.1.1'
+release = '0.2.0 '
 
 
 # -- General configuration ---------------------------------------------------