+++ examples/example.py
@@ -2,7 +2,7 @@
 # create a vector table, 
 # insert 10 vectors, 
 # and execute a vector similarity search.
-from milvus import Milvus, Prepare, IndexType, Status
+from milvus import Milvus, IndexType, MetricType
 import time
 
 # Milvus server IP address and port.
@@ -26,8 +26,8 @@ def main():
         param = {
             'table_name': table_name,
             'dimension': 16,
-            'index_type': IndexType.FLAT,
-            'store_raw_vector': False
+            'index_file_size': 1024,
+            'metric_type': MetricType.L2
         }
 
         milvus.create_table(param)
@@ -58,7 +58,7 @@ def main():
     # Wait for 6 seconds, since Milvus server persist vector data every 5 seconds by default.
     # You can set data persist interval in Milvus config file.
     time.sleep(6)
-    
+
     # Get demo_table row count
     status, result = milvus.get_table_row_count(table_name)
 
@@ -72,11 +72,12 @@ def main():
         'table_name': table_name,
         'query_records': query_vectors,
         'top_k': 1,
+        'nprobe': 16
 
     }
     status, results = milvus.search_vectors(**param)
 
-    if results[0][0].distance== 0.0 or results[0][0].id == ids[3]:
+    if results[0][0].distance == 0.0 or results[0][0].id == ids[3]:
         print('Query result is correct')
     else:
         print('Query result isn\'t correct')