+++ examples/simple_example.py
@@ -1,6 +1,8 @@
 # This program demos how to create a Milvus table, insert 20 vectors and get the table row count.
-from milvus import Milvus, Prepare, IndexType, Status
-import random, time
+from milvus import Milvus, IndexType, MetricType
+
+import random
+import time
 
 milvus = Milvus()
 
@@ -13,10 +15,10 @@ table_name = 'demo_table_02'
 
 # Create table: table name, vector dimension and index type
 if not milvus.has_table(table_name):
-    milvus.create_table({'table_name': table_name, 'dimension': 256, 'index_type': IndexType.FLAT})
+    milvus.create_table({'table_name': table_name, 'dimension': 256, 'index_file_size': 1024, 'metric_type': MetricType.L2})
 
 # Insert 20 256-dim-vectors into demo_table
-vectors = [[random.random()for _ in range(256)] for _ in range(20)]
+vectors = [[random.random() for _ in range(256)] for _ in range(20)]
 milvus.add_vectors(table_name=table_name, records=vectors)
 time.sleep(1)