+++ milvus/__init__.py
@@ -1,6 +1,7 @@
-from .client.Client import Milvus, Prepare, Status, IndexType
+from .client.GrpcClient import GrpcMilvus as Milvus
+from .client.GrpcClient import Prepare
+from .client.types import IndexType, MetricType, Status
+from .client import Exceptions as milvusError
+from .client import __version__
 
-__all__ = ['Milvus', 'Prepare', 'Status', 'IndexType', '__version__']
-
-__version__ = '0.1.25'
-
+__all__ = ['Milvus', 'Prepare', 'Status', 'IndexType', 'MetricType', 'milvusError', '__version__']