+++ tests/test_thread.py
@@ -1,86 +1,59 @@
 import sys
-sys.path.append('.')
-from milvus import Milvus, IndexType
-from pprint import pprint
-import concurrent.futures
-import threading
 import random
-import os
+import time
+import concurrent.futures
 
-dimension = 256
-number = 10000
-pool_size = 1000
-table_pool_size = 1000
-step = number//pool_size
+sys.path.append('.')
+from factorys import time_it
 
-vectors = [[random.random()for _ in range(dimension)] for _ in range(number)]
+dimension = 512
+number = 100000
 table_name = 'multi_task'
 
-start_count = 0
-count = 0
 
 def add_vector_task(milvus, vector):
-    global count
-    global start_count
-    start_count += 1
-    print('start....................{}'.format(start_count))
     status, ids = milvus.add_vectors(table_name=table_name, records=vector)
-    count += 1
-    print("end...............{}".format(count))
+
+    assert status.OK(), "add vectors failed"
+    assert len(ids) == len(vector)
 
 
-def thread_pool_add_vector():
-
-    milvus = Milvus()
-    milvus.connect(uri='tcp://127.0.0.1:19530')
-
-    if not milvus.has_table(table_name):
-        param = {
-            'table_name': table_name,
-            'dimension': dimension,
-            'index_type': IndexType.FLAT,
-            'store_raw_vector': False
-        }
-
-        milvus.create_table(param)
-
+@time_it
+def grpc_thread_pool_add_vector(milvus, pool_size, vectors):
     with concurrent.futures.ThreadPoolExecutor(max_workers=pool_size) as executor:
-        for x in range(0, number, step):
-            executor.submit(add_vector_task, milvus, vectors[x:x+step])
+        for x in range(pool_size):
+            executor.submit(add_vector_task, milvus, vectors)
 
 
-def add_table_task(milvus):
-    global count
-    global start_count
-    start_count += 1
-    param = {
-        'table_name': unique_table_name(),
+def test_grpc_run(gcon):
+    gmilvus = gcon
+
+    if gmilvus is None:
+        assert False, "Error occurred: connect failure"
+
+    if gmilvus.has_table(table_name):
+        gmilvus.delete_table(table_name)
+        time.sleep(2)
+
+    table_param = {
+        'table_name': table_name,
         'dimension': dimension,
-        'index_type': IndexType.FLAT,
-        'store_raw_vector': False
+        'index_file_size': 1024,
+        'metric_type': 1
     }
-    print('start....................{}'.format(start_count))
-    status = milvus.create_table(param)
-    count += 1
-    print("end...............{}".format(count))
 
-def unique_table_name():
-    return str(random.random())
+    gmilvus.create_table(table_param)
 
-def thread_pool_add_table():
-    milvus = Milvus()
-    milvus.connect(uri='tcp://localhost:19530')
+    for p in (1, 5, 10, 20, 40, 50, 100):
+        pool_size = p
+    step = number // pool_size
 
-    with concurrent.futures.ThreadPoolExecutor(max_workers=table_pool_size) as executor:
-        for x in range(0, table_pool_size):
+    vectors = [[random.random() for _ in range(dimension)] for _ in range(step)]
 
-            executor.submit(add_table_task, milvus)
+    print(f'Thread pool size................... {p}')
+    grpc_thread_pool_add_vector(gmilvus, p, vectors)
 
-
-
-if __name__ == '__main__':
-    thread_pool_add_vector()
-    #thread_pool_add_table()
-    
-
+    time.sleep(1.5)
+    _, gcount = gmilvus.get_table_row_count(table_name)
+    print(gcount)