+++ CHANGELOG.md
@@ -23,6 +23,19 @@
 - \#59 ISSUE: **[MS-182]** fix query ranges param not control bug
 - \#60 ISSUE: **[MS-185]** fix connection no timeout bug 
 - \#69 ISSUE: **[MS-243]** fixed
+- \#83 ISSUE: fix celery server create table error
+- \#   ISSUE:  
+    - **[MS-424]** IndexParam str more friendly
+    - **[MS-433]** fix wrong passing param bug
+    - **[MS-438]** fix add vectors with ids bug
+    - **[MS-439]** fix search result unpack bug on windows
+    - **[MS-452]** fix build index timeout bug
+    - **[MS-444]** fix timeout bug
+    - **[MS-469]** add index_file_size min value check
+    - **[MS-512]** fix ids error
+    - **[MS-521]** fix ids check error
+    - **[MS-522]** fix index_param check error
+    
 ### Improvement
 ---
 - \#10 ISSUE: Update examples
@@ -33,6 +46,9 @@
 - \#45, \#47 ISSUE:  Update AdvancedExample
 - \#51 ISSUE: Brute force thread_safe for sdk
 - \#71 ISSUE: Update thrift api, search vectors results are binaries
+- \#73 ISSUE: Change table schema print format
+- \#74 ISSUE: Add a new attribute server_address
+- \#77 ISSUE: print out TopKResult more friendly
 ### New Feature
 ---
 - \#3
@@ -60,9 +76,20 @@
     - support search by range
     - fix server_status return None bug
     
+- \#76 ISSUE
+    - replace thrift with grpc
+
 - \#79 ISSUE
     - add new indextype MIN_NSG
     - fix indextype name bug
+
+- \#81 ISSUE
+    - remove all thrift 
+    - update interface and interface param
+ 
+- \#82
+    - add timeout option
+    - add default param in tableschema and indexparam
 ### Task
 ---
 - \#1 Build Repository